requirements.txt
fastapi
uvicorn
sqlalchemy
postgresql
tritonclient
anthropic
kafka-python